try to fix the long long warnings

+++ src/Matrix.cpp
@@ -40,9 +40,9 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
     this->granularity = 1.0;
   }
   
-  matInt = new long long *[length];
+  matInt = new qlonglong *[length];
   for (int k = 0; k < 4; k++ ) {
-    matInt[k] = new long long[length];
+    matInt[k] = new qlonglong[length];
     for (int p = 0 ; p < length; p++) {
       matInt[k][p] = ROUND_TO_INT((double)(mat[k][p]*this->granularity)); 
     }
@@ -77,14 +77,14 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
   
   if (sortColumns) {
     // sort the columns : the first column is the one with the greatest value
-    long long min = 0;
+    qlonglong min = 0;
     for (int i = 0; i < length; i++) {
       for (int k = 0; k < 4; k++) {
         min = MIN(min,matInt[k][i]);
       }
     }
     min --;
-    long long *maxs = new long long [length];
+    qlonglong *maxs = new qlonglong [length];
     for (int i = 0; i < length; i++) {
       maxs[i] = matInt[0][i];
       for (int k = 1; k < 4; k++) {
@@ -93,12 +93,12 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
         }
       }
     }
-    long long **mattemp = new long long *[4];
+    qlonglong **mattemp = new qlonglong *[4];
     for (int k = 0; k < 4; k++) {        
-      mattemp[k] = new long long [length];
+      mattemp[k] = new qlonglong [length];
     }
     for (int i = 0; i < length; i++) {
-      long long max = maxs[0];
+      qlonglong max = maxs[0];
       int p = 0;
       for (int j = 1; j < length; j++) {
         if (max < maxs[j]) {
@@ -135,9 +135,9 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
   
   // computes offsets
   this->offset = 0;
-  offsets = new long long [length];
+  offsets = new qlonglong [length];
   for (int i = 0; i < length; i++) {
-    long long min = matInt[0][i];
+    qlonglong min = matInt[0][i];
     for (int k = 1; k < 4; k++ )  {
       min = ((min < matInt[k][i])?min:(matInt[k][i]));
     }
@@ -164,9 +164,9 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
   
   
   // look for the minimum score of the matrix for each column
-  minScoreColumn = new long long [length];
-  maxScoreColumn = new long long [length];
-  sum            = new long long [length];
+  minScoreColumn = new qlonglong [length];
+  maxScoreColumn = new qlonglong [length];
+  sum            = new qlonglong [length];
   minScore = 0;
   maxScore = 0;
   for (int i = 0; i < length; i++) {
@@ -193,8 +193,8 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
   //cout << "SCORE RANGE : " << minScore << " - " << maxScore << " : " << this->scoreRange << endl;
 #endif
   
-  bestScore = new long long[length];
-  worstScore = new long long[length];
+  bestScore = new qlonglong[length];
+  worstScore = new qlonglong[length];
   bestScore[length-1] = maxScore;
   worstScore[length-1] = minScore;
   for (int i = length - 2; i >= 0; i--) {
@@ -210,18 +210,18 @@ void Matrix::computesIntegerMatrix (double granularity, bool sortColumns) {
 /**
 * Computes the pvalue associated with the threshold score requestedScore.
  */
-void Matrix::lookForPvalue (long long requestedScore, long long min, long long max, double *pmin, double *pmax) {
+void Matrix::lookForPvalue (qlonglong requestedScore, qlonglong min, qlonglong max, double *pmin, double *pmax) {
   
-  map<long long, double> *nbocc = calcDistribWithMapMinMax(min,max); 
-  map<long long, double>::iterator iter;
+  map<qlonglong, double> *nbocc = calcDistribWithMapMinMax(min,max); 
+  map<qlonglong, double>::iterator iter;
   
 #ifdef SHOWCERR
   //cerr << "  Looks for Pvalue between " << min << " and " << max << " for score " << requestedScore << endl;
 #endif
   // computes p values and stores them in nbocc[length] 
   double sum = nbocc[length][max+1];
-  long long s = max + 1;
-  map<long long, double>::reverse_iterator riter = nbocc[length-1].rbegin();
+  qlonglong s = max + 1;
+  map<qlonglong, double>::reverse_iterator riter = nbocc[length-1].rbegin();
   while (riter != nbocc[length-1].rend()) {
     sum += riter->second;
     if (riter->first >= requestedScore) s = riter->first;
@@ -255,18 +255,18 @@ void Matrix::lookForPvalue (long long requestedScore, long long min, long long m
 /**
 * Computes the score associated with the pvalue requestedPvalue.
  */
-long long Matrix::lookForScore (long long min, long long max, double requestedPvalue, double *rpv, double *rppv) {
+qlonglong Matrix::lookForScore (qlonglong min, qlonglong max, double requestedPvalue, double *rpv, double *rppv) {
   
-  map<long long, double> *nbocc = calcDistribWithMapMinMax(min,max); 
-  map<long long, double>::iterator iter;
+  map<qlonglong, double> *nbocc = calcDistribWithMapMinMax(min,max); 
+  map<qlonglong, double>::iterator iter;
 #ifdef SHOWCERR
   //cerr << "  Looks for score between " << min << " and " << max << endl;
 #endif
   // computes p values and stores them in nbocc[length] 
   double sum = 0.0;
-  map<long long, double>::reverse_iterator riter = nbocc[length-1].rbegin();
-  long long alpha = riter->first+1;
-  long long alpha_E = alpha;
+  map<qlonglong, double>::reverse_iterator riter = nbocc[length-1].rbegin();
+  qlonglong alpha = riter->first+1;
+  qlonglong alpha_E = alpha;
   nbocc[length][alpha] = 0.0;
   while (riter != nbocc[length-1].rend()) {
     sum += riter->second;
@@ -327,15 +327,15 @@ long long Matrix::lookForScore (long long min, long long max, double requestedPv
 
 // computes the distribution of scores between score min and max as the DP algrithm proceeds 
 // but instead of using a table we use a map to avoid computations for scores that cannot be reached
-map<long long, double> *Matrix::calcDistribWithMapMinMax (long long min, long long max) { 
+map<qlonglong, double> *Matrix::calcDistribWithMapMinMax (qlonglong min, qlonglong max) { 
   
   // maps for each step of the computation
   // nbocc[length] stores the pvalue
   // nbocc[pos] for pos < length stores the qvalue
-  map<long long, double> *nbocc = new map<long long, double> [length+1];
-  map<long long, double>::iterator iter;
+  map<qlonglong, double> *nbocc = new map<qlonglong, double> [length+1];
+  map<qlonglong, double>::iterator iter;
   
-  long long *maxs = new long long[length+1]; // @ pos i maximum score reachable with the suffix matrix from i to length-1
+  qlonglong *maxs = new qlonglong[length+1]; // @ pos i maximum score reachable with the suffix matrix from i to length-1
   
 #ifdef VERBOSE    
   //cerr << "  Calc distrib between " << min << " and " << max << endl;
@@ -359,7 +359,7 @@ map<long long, double> *Matrix::calcDistribWithMapMinMax (long long min, long lo
     iter = nbocc[pos-1].begin();
     while (iter != nbocc[pos-1].end()) {
       for (int k = 0; k < 4; k++) {
-        long long sc = iter->first + matInt[k][pos];
+        qlonglong sc = iter->first + matInt[k][pos];
         if (sc+maxs[pos+1] >= min) {
           // the score min can be reached
           if (sc > max) {
@@ -388,15 +388,15 @@ map<long long, double> *Matrix::calcDistribWithMapMinMax (long long min, long lo
 
 
 
-long long Matrix::fastPvalue (Matrix *m, long long alpha) {
+qlonglong Matrix::fastPvalue (Matrix *m, qlonglong alpha) {
   
   
-  map<long long, long long> *q = new map<long long, long long> [m->length+1];
-  map<long long, long long>::iterator iter;
+  map<qlonglong, qlonglong> *q = new map<qlonglong, qlonglong> [m->length+1];
+  map<qlonglong, qlonglong>::iterator iter;
   
-  long long P = 0;
+  qlonglong P = 0;
   
-  long long *maxm = new long long[m->length+1]; // @ pos i maximum score reachable with the suffix matrix from i to length-1
+  qlonglong *maxm = new qlonglong[m->length+1]; // @ pos i maximum score reachable with the suffix matrix from i to length-1
   
   maxm[m->length] = 0;
   for (int i = m->length-1; i >= 0; i--) {
@@ -417,10 +417,10 @@ long long Matrix::fastPvalue (Matrix *m, long long alpha) {
     iter = q[pos-1].begin();
     while (iter != q[pos-1].end()) {
       for (int k = 0; k < 4; k++) {
-        long long scm = iter->first + m->matInt[k][pos];
+        qlonglong scm = iter->first + m->matInt[k][pos];
         if (scm > alpha) {
           //cout << "Update P from " << P;
-          P += iter->second * (long long)pow(4.0,m->length-pos-1); 
+          P += iter->second * (qlonglong)pow(4.0,m->length-pos-1); 
           //cout << " to P " << P << endl;
         } else if (scm + maxm[pos+1] > alpha) {
           q[pos][scm] += iter->second;
